+++ sdk/client_test.go
@@ -33,8 +33,11 @@ var (
 	//go:embed testdata/capa/v25.1.0-demo.0/release.yaml
 	v2510ReleaseManifest string
 
-	//go:embed testdata/capa/github_get_capa_for_git.json
-	capaDirectoryResponse string
+	//go:embed testdata/capa/github_get_capa_for_git_default.json
+	capaDirectoryGitDefaultResponse string
+
+	//go:embed testdata/capa/github_get_capa_for_git_ref.json
+	capaDirectoryGitReferenceResponse string
 
 	//go:embed testdata/capa/v25.0.0/release_yaml_content_response.json
 	v25ReleaseManifestContentResponse string
@@ -143,6 +146,22 @@ var _ = Describe("Client", func() {
 		}
 	})
 
+	It("Gets new releases for the specified provider and git reference", func() {
+		ctx := context.Background()
+		releases, err := releasesClient.GetNewReleasesForGitReference(ctx, ProviderAws, "add-capa-v25.0.0")
+		Expect(err).NotTo(HaveOccurred())
+
+		expectedReleases := []string{
+			"25.0.0",
+		}
+
+		for i, release := range releases {
+			resultReleaseVersion, err := release.GetVersion()
+			Expect(err).NotTo(HaveOccurred())
+			Expect(resultReleaseVersion).To(Equal(expectedReleases[i]))
+		}
+	})
+
 	AfterEach(func() {
 		server.Close()
 	})
@@ -172,7 +191,18 @@ func createAndStartTestServer() *httptest.Server {
 	})
 	mux.HandleFunc("/api/v3/repos/giantswarm/releases/contents/capa", func(rw http.ResponseWriter, req *http.Request) {
 		// return GitHub response JSON
-		_, err := rw.Write([]byte(capaDirectoryResponse))
+		// Get the query parameters
+		queryParams := req.URL.Query()
+		// Retrieve specific query parameter
+		ref := queryParams.Get("ref")
+		var err error
+
+		switch ref {
+		case "add-capa-v25.0.0":
+			_, err = rw.Write([]byte(capaDirectoryGitReferenceResponse))
+		default:
+			_, err = rw.Write([]byte(capaDirectoryGitDefaultResponse))
+		}
 		Expect(err).NotTo(HaveOccurred())
 	})
 	mux.HandleFunc("/api/v3/repos/giantswarm/releases/contents/capa/v25.0.0/release.yaml", func(rw http.ResponseWriter, req *http.Request) {